Infinite scroll bug is fixed
Instead of converting the future data into interable inside the Future
Builder, i just did in a separate async function

The problem with FutureBuilder is that whenever it was getting new data
it was building the entire widget from scratch

With GridView.builder() as it already only keeps the part of the list in
memory that is visible in viewport, adding things to the list doesnt
rebuild the entire widget and as a result i dont get thrown back to the
top after each api request

Next tasks
	- Deal with edge cases such as when images cant be fetched due
	  to an error.
	- No more images available edge case
	- Add a spinner during loading
Another bug with the current version of this app is that the
scrollController will continuosly send request to Pexels as many times
as the user hits the end of the list.
To prevent this we need to add isLoading boolean to make sure request is
only send once.